rfp_analyser/requirements.txt
# Build dependencies
wheel>=0.42.0
setuptools>=68.0.0
build>=1.0.3

# Core dependencies
pdfminer.six>=20221105
# pytesseract>=0.3.10
Pillow>=10.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
chromadb>=0.4.24
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.2
python-multipart>=0.0.6
streamlit>=1.31.0
reportlab>=4.0.8
python-pptx>=0.6.23
requests>=2.31.0
ollama>=0.1.6
paddleocr>=2.7.0
PyMuPDF>=1.23.0  # for PDF processing

# Pre-built wheels
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Additional dependencies
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0 